What I should have done the first time around ...
