+++ InfoTrack.CachingTokenClient.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 16.0.28803.202
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "InfoTrack.OAuth", "src\InfoTrack.OAuth\InfoTrack.OAuth.csproj", "{E3CD85B5-8EDE-4B0C-9C40-99CA0A9F778C}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "InfoTrack.OAuth.Caching.DotNetCore", "src\InfoTrack.OAuth.Caching.DotNetCore\InfoTrack.OAuth.Caching.DotNetCore.csproj", "{AB045B23-FC55-46B1-AC2D-904F1D1FC832}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "InfoTrack.OAuth.Caching.DotNetFramework", "src\InfoTrack.OAuth.Caching.DotNetFramework\InfoTrack.OAuth.Caching.DotNetFramework.csproj", "{42B5F2F7-3A22-4EDB-A4A1-666C170F340F}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{1F49F9A6-6A36-4597-B419-CE8812B376AB}"
@@ -20,7 +18,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		readme.md = readme.md
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InfoTrack.OAuth.Caching.DotNetStandard", "src\InfoTrack.OAuth.Caching.DotNetStandard\InfoTrack.OAuth.Caching.DotNetStandard.csproj", "{F322BD3D-2EA1-49A8-9EAC-BE8083261E5E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "InfoTrack.OAuth.Caching.DotNetStandard", "src\InfoTrack.OAuth.Caching.DotNetStandard\InfoTrack.OAuth.Caching.DotNetStandard.csproj", "{F322BD3D-2EA1-49A8-9EAC-BE8083261E5E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -32,10 +30,6 @@ Global
 		{E3CD85B5-8EDE-4B0C-9C40-99CA0A9F778C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E3CD85B5-8EDE-4B0C-9C40-99CA0A9F778C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E3CD85B5-8EDE-4B0C-9C40-99CA0A9F778C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AB045B23-FC55-46B1-AC2D-904F1D1FC832}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AB045B23-FC55-46B1-AC2D-904F1D1FC832}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AB045B23-FC55-46B1-AC2D-904F1D1FC832}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AB045B23-FC55-46B1-AC2D-904F1D1FC832}.Release|Any CPU.Build.0 = Release|Any CPU
 		{42B5F2F7-3A22-4EDB-A4A1-666C170F340F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{42B5F2F7-3A22-4EDB-A4A1-666C170F340F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{42B5F2F7-3A22-4EDB-A4A1-666C170F340F}.Release|Any CPU.ActiveCfg = Release|Any CPU
