+++ src/InfoTrack.OAuth.Caching.DotNetStandard/CachingTokenClient.cs
@@ -1,61 +1,33 @@
 ﻿using System;
-using System.Runtime.Caching;
-using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.Extensions.Caching.Memory;
 
 namespace InfoTrack.OAuth.Caching.DotNetStandard
 {
     public class CachingTokenClient : BaseCachingTokenClient
     {
-        private static readonly SemaphoreSlim _semaphore = new SemaphoreSlim(1, 1);
-        private readonly ObjectCache _cache;
+        private readonly IMemoryCache _memoryCache;
 
-        public CachingTokenClient()
-            : this(MemoryCache.Default) { }
+        public CachingTokenClient(IMemoryCache memoryCache)
+            : this(memoryCache, ClientOptions.Default) { }
 
-        public CachingTokenClient(ClientOptions clientOptions)
-            : this(MemoryCache.Default, clientOptions) { }
-
-        public CachingTokenClient(ObjectCache cache)
-            : this(cache, ClientOptions.Default) { }
-
-        public CachingTokenClient(ObjectCache cache, ClientOptions clientOptions)
+        public CachingTokenClient(IMemoryCache memoryCache, ClientOptions clientOptions)
             : base(clientOptions)
         {
-            _cache = cache ?? throw new ArgumentNullException(nameof(cache));
+            _memoryCache = memoryCache ?? throw new ArgumentNullException(nameof(memoryCache));
         }
 
         protected override async Task<TItem> GetOrCreateAsync<TItem>(string key, Func<CacheItem, Task<TItem>> factory)
         {
-            var item = _cache.Get(key);
-
-            if (item != null)
+            return await _memoryCache.GetOrCreateAsync(key, async cacheEntry =>
             {
-                return (TItem)item;
-            }
-
-            await _semaphore.WaitAsync();
-
-            try
-            {
-                item = _cache.Get(key);
-
-                if (item != null)
-                {
-                    return (TItem)item;
-                }
-
                 var cacheItem = new CacheItem();
-                item = await factory(cacheItem);
+                var item = await factory(cacheItem);
 
-                _cache.Set(key, item, new CacheItemPolicy() { AbsoluteExpiration = cacheItem.AbsoluteExpiration });
+                cacheEntry.AbsoluteExpiration = cacheItem.AbsoluteExpiration;
 
-                return (TItem)item;
-            }
-            finally
-            {
-                _semaphore.Release();
-            }
+                return item;
+            });
         }
     }
 }
