+++ src/InfoTrack.OAuth.Caching.DotNetStandard/InfoTrack.OAuth.Caching.DotNetStandard.csproj
@@ -1,7 +1,16 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
+    <PackageLicenseExpression></PackageLicenseExpression>
+    <PackageLicenseFile>LICENSE.txt</PackageLicenseFile>
+    <Authors>InfoTrack</Authors>
+    <Product>InfoTrack.CachingTokenClient</Product>
+    <Description>OAuth 2 Token Client which automatically caches the authentication server response.</Description>
+    <PackageProjectUrl>https://github.com/InfoTrackGlobal/CachingTokenClient</PackageProjectUrl>
+    <PackageTags>oauth</PackageTags>
   </PropertyGroup>
 
   <ItemGroup>
@@ -12,13 +21,11 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\InfoTrack.OAuth\InfoTrack.OAuth.csproj" />
+    <PackageReference Include="Microsoft.Extensions.Caching.Abstractions" Version="1.1.2" />
   </ItemGroup>
 
   <ItemGroup>
-    <Reference Include="System.Runtime.Caching">
-      <HintPath>..\..\..\..\Program Files (x86)\Reference Assemblies\Microsoft\Framework\.NETFramework\v4.5.2\System.Runtime.Caching.dll</HintPath>
-    </Reference>
+    <ProjectReference Include="..\InfoTrack.OAuth\InfoTrack.OAuth.csproj" />
   </ItemGroup>
 
 </Project>
